+++ src/main/java/io/github/grrolland/hcshm/commands/TouchCommand.java
@@ -1,17 +1,17 @@
 /**
  * ngx-distributed-shm
  * Copyright (C) 2018  Flu.Tech
- *
+ * <p>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- *
+ * <p>
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- *
+ * <p>
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
@@ -40,16 +40,13 @@ public class TouchCommand extends Command {
      */
     public String execute(String[] commandTokens) {
         final StringBuilder response = new StringBuilder();
-        try
-        {
+        try {
             assertTokens(commandTokens, 3);
             String key = getKey(commandTokens[1]);
-            int expire = getExpire(commandTokens[2]);
+            long expire = getExpire(commandTokens[2]);
             getService().touch(key, expire);
             writeDone(response);
-        }
-        catch (ProtocolException e)
-        {
+        } catch (ProtocolException e) {
             writeMalformedRequest(response);
         }
         return response.toString();
